Update test for SelectEventTags and rename to SelectTags
