+++ src/pages/createOrEditTask/selectors/index.js
@@ -1,1 +1,1 @@
-export { default as selectEventTags } from "./selectEventTags";
+export { default as selectTags } from "./selectTags";
